requirements.txt
Brotli==1.1.0
cssselect2==0.8.0
fonttools==4.58.5
Markdown==3.8.2
pillow==11.3.0
pydyf==0.11.0
pyphen==0.17.2
python-dotenv==1.1.1
tinyhtml5==2.0.0
weasyprint==65.1
zopfli==0.2.3.post1
